requirements.txt
flask>=2.0.1
flask-login>=0.5.0
werkzeug>=2.0.1
numpy==1.23.5
pandas==1.5.3
loguru>=0.5.3
python-dotenv>=0.19.0
gunicorn>=20.1.0
ibapi>=9.81.1 